# src/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0",
        "description": "Documentation for your API"
    },
    "servers": [
        {
            "url": "http://localhost:3333" 
        }
    ],
    "paths": {
      "/api/tarefa/":{
        "get":{
          "summary":"Get all Tarefas",
          "tags": ["Tarefas"],
          "security":[{"bearerAuth":[]}],
          "responses":{
            "401":{
              "description":"Unauthorized"
            },
            "400":{
              "description":"Erro na consulta"
            },
            "200":{
              "description":"OK",
              "content":{
                "aplication/json":{
                  "schema":{
                    "type": "array",
                    "items":{
                      "$ref":"#/components/schemas/Tarefa"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/tarefa/byIdTarefa":{
        "get":{
          "summary":"busca por idTarefa",
          "tags": ["Tarefas"],
          "security":[{"bearerAuth":[]}],
          "parameters": [
            {
              "required":true,
              "name": "idTar",
              "in": "query",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses":{
            "401":{
              "description":"Unauthorized"
            },
            "400":{
              "description":"Erro na consulta"
            },
            "200":{
              "description":"OK",
              "content":{
                "aplication/json":{
                  "schema":{
                    "type": "array",
                    "items":{
                      "$ref":"#/components/schemas/Tarefa"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/tarefa/ByNome":{
        "get":{
          "summary":"busca por nome",
          "tags": ["Tarefas"],
          "security":[{"bearerAuth":[]}],
          "parameters": [
            {
              "required":true,
              "name": "nmTar",
              "in": "query",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses":{
            "401":{
              "description":"Unauthorized"
            },
            "400":{
              "description":"Erro na consulta"
            },
            "200":{
              "description":"OK",
              "content":{
                "aplication/json":{
                  "schema":{
                    "type": "array",
                    "items":{
                      "$ref":"#/components/schemas/Tarefa"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/tarefa/ByidUsuario":{
        "get":{
          "summary":"Buscar tarefa por Id_Usuario",
          "tags": ["Tarefas"],
          "security":[{"bearerAuth":[]}],
          "parameters": [
            {
              "required":true,
              "name": "idusr",
              "in": "query",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses":{
            "401":{
              "description":"Unauthorized"
            },
            "400":{
              "description":"Registro Inexistente"
            },
            "200":{
              "description":"OK",
              "content":{
                "aplication/json":{
                  "schema":{
                    "type": "array",
                    "items":{
                      "$ref":"#/components/schemas/Tarefa"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/tarefa/dtTarefa":{
        "get":{
          "summary":"Buscar tarefa por periodo",
          "tags": ["Tarefas"],
          "security":[{"bearerAuth":[]}],
          "parameters": [
            {
              "description":"yyyy-mm-dd",
              "required":true, 
              "name": "dtStart",
              "in": "query",
              "schema": {
                "type": "string"
              }
            },
            {
              "description":"yyyy-mm-dd",
              "required":true,
              "name": "dtEnd",
              "in": "query",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses":{
            "401":{
              "description":"Unauthorized"
            },
            "400":{
              "description":"Erro na consulta"
            },
            "200":{
              "description":"OK",
              "content":{
                "aplication/json":{
                  "schema":{
                    "type": "array",
                    "items":{
                      "$ref":"#/components/schemas/Tarefa"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/tarefa/weekUsaer":{
        "get":{
          "summary":"Buscar tarefa por usuario e semana",
          "tags": ["Tarefas"],
          "security":[{"bearerAuth":[]}],
          "parameters": [
            {
              "required":true, 
              "name": "idusuario",
              "in": "query",
              "schema": {
                "type": "string"
              }
            },
            {
              "required":true,
              "name": "nrsemana",
              "in": "query",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses":{
            "401":{
              "description":"Unauthorized"
            },
            "400":{
              "description":"Erro na consulta"
            },
            "200":{
              "description":"OK",
              "content":{
                "aplication/json":{
                  "schema":{
                    "type": "array",
                    "items":{
                      "$ref":"#/components/schemas/Tarefa"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/tarefa":{
        "post":{
          "summary":"Insere novo usuario",
          "tags": ["Tarefas"],
          "security":[{"bearerAuth":[]}],
          "requestBody":{
            "description":"",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/Tarefa"
                }
              },
              "text/json": {
                "schema": {
                  "$ref":"#/components/schemas/Tarefa"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref":"#/components/schemas/Tarefa"
                }
              }
            }
          },
          "responses":{
            "401":{
              "description":"Unauthorized"
            },
            "400":{
              "description":"Tarefa ja existe."
            },
            "200":{
              "description":"OK",
              "content":{
                "aplication/json":{
                  "schema":{
                    "type": "object",
                    "$ref":"#/components/schemas/Tarefa"                      
                  }
                }
              }
            }
          }
        }
      },
      "/api/tarefa/dell":{
        "delete":{
          "summary":"Deleta registro Tarefa",
          "tags": ["Tarefas"],
          "security":[{"bearerAuth":[]}],
          "parameters": [
            {
              "required":true,
              "name": "idTarefa",
              "in": "query",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses":{
            "401":{
              "description":"Unauthorized"
            },
            "400":{
              "description":"Erro na consulta"
            },
            "200":{
              "description":"OK",
              "content":{
                "aplication/json":{
                  "schema":{
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/api/jwtValidator":{
        "post":{
          "summary":"Valida JWT",
          "tags": ["Authentication"],
          "requestBody":{
            "description":"",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/jwt"
                }
              },
              "text/json": {
                "schema": {
                  "$ref":"#/components/schemas/jwt"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref":"#/components/schemas/jwt"
                }
              }
            }
          },
          "responses":{
            "400":{
                "description":"Falha na validacao"
              },
            "200":{
              "description":"OK",
              "content":{
                "aplication/json":{
                  "schema":{
                    "type": "object",
                      "$ref":"#/components/schemas/Retjwt"
                  }
                }
              }
            }
          }
        }
      },
      "/api/login":{
        "post":{
          "summary":"Gera JWT",
          "tags": ["Authentication"],
          "requestBody":{
            "description":"",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/Login"
                }
              },
              "text/json": {
                "schema": {
                  "$ref":"#/components/schemas/Login"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref":"#/components/schemas/Login"
                }
              }
            }
          },
          "responses":{
            "400":{
                "description":"Falha na autenticação"
              },
            "200":{
              "description":"OK",
              "content":{
                "aplication/json":{
                  "schema":{
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/api/users/":{
        "get":{
          "summary":"Get all users",
          "tags": ["Users"],
          "security":[{"bearerAuth":[]}],
          "responses":{
            "401":{
              "description":"Unauthorized"
            },
            "400":{
              "description":"Usuario ja existe."
            },
            "200":{
              "description":"OK",
              "content":{
                "aplication/json":{
                  "schema":{
                    "type": "array",
                    "items":{
                      "$ref":"#/components/schemas/Users"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/users/by":{
        "get":{
          "summary":"busca por nome, email ou password",
          "tags": ["Users"],
          "security":[{"bearerAuth":[]}],
          "parameters": [
            {
              "required":true,
              "name": "param",
              "in": "query",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses":{
            "401":{
              "description":"Unauthorized"
            },
            "400":{
              "description":"Usuario ja existe."
            },
            "200":{
              "description":"OK",
              "content":{
                "aplication/json":{
                  "schema":{
                    "type": "object",
                    "$ref":"#/components/schemas/Users"
                  }
                }
              }
            }
          }
        }
      },
      "/api/users":{
        "post":{
          "summary":"Insere novo usuario",
          "tags": ["Users"],
          "security":[{"bearerAuth":[]}],
          "requestBody":{
            "description":"",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/Users"
                }
              },
              "text/json": {
                "schema": {
                  "$ref":"#/components/schemas/Users"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref":"#/components/schemas/Users"
                }
              }
            }
          },
          "responses":{
            "401":{
              "description":"Unauthorized"
            },
            "400":{
              "description":"Usuario ja existe."
            },
            "200":{
              "description":"OK",
              "content":{
                "aplication/json":{
                  "schema":{
                    "type": "object",
                    "$ref":"#/components/schemas/Users"                      
                  }
                }
              }
            }
          }
        }
      },
      "/api/users/recuperacao":{
        "post":{
          "summary":"Recupera usuario e senha",
          "tags": ["Users"],
          
          "requestBody":{
            "description":"",
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties": {
                    "email":{
                      "type":"string"
                    }
                  }
                }
              },
              "text/json": {
                "schema":{
                  "type":"object",
                  "properties": {
                    "email":{
                      "type":"string"
                    }
                  }
                }
              },
              "application/*+json": {
                "schema":{
                  "type":"object",
                  "properties": {
                    "statemailus":{
                      "type":"string"
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "401":{
              "description":"Unauthorized"
            },
            "400":{
              "description":"Bad Request"
            },
            "200":{
              "description":"OK",
              "content":{
                "aplication/json":{
                  "schema":{
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/api/email":{
        "post":{
          "summary":"Envia Emails",
          "tags": ["Emails"],
          "security":[{"bearerAuth":[]}],
          "requestBody":{
            "description":"",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/SendMail"
                }
              },
              "text/json": {
                "schema": {
                  "$ref":"#/components/schemas/SendMail"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref":"#/components/schemas/SendMail"
                }
              }
            }
          },
          "responses":{
            "401":{
              "description":"Unauthorized"
            },
            "400":{
              "description":"Usuario ja existe."
            },
            "200":{
              "description":"OK",
              "content":{
                "aplication/json":{
                  "schema":{
                    "type": "string"                   
                  }
                }
              }
            }
          }
        }
      },
      "/api/tpTarefa/":{
        "get":{
          "summary":"Get all Tipo Tarefas",
          "tags": ["Tipo Tarefa"],
          "security":[{"bearerAuth":[]}],
          "responses":{
            "401":{
              "description":"Unauthorized"
            },
            "400":{
              "description":"ja existe."
            },
            "200":{
              "description":"OK",
              "content":{
                "aplication/json":{
                  "schema":{
                    "type": "array",
                    "items":{
                      "$ref":"#/components/schemas/TipoTarefa"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/tpTarefa/by":{
        "get":{
          "summary":"Get Tarefas por prametro",
          "tags": ["Tipo Tarefa"],
          "security":[{"bearerAuth":[]}],
          "parameters": [
            {
              "required":true,
              "name": "param",
              "in": "query",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses":{
            "401":{
              "description":"Unauthorized"
            },
            "400":{
              "description":"Erro."
            },
            "200":{
              "description":"OK",
              "content":{
                "aplication/json":{
                  "schema":{
                    "type": "array",
                    "items":{
                      "$ref":"#/components/schemas/TipoTarefa"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/tpTarefa":{
        "post":{
          "summary":"Insere novo usuario",
          "tags": ["Tipo Tarefa"],
          "security":[{"bearerAuth":[]}],
          "requestBody":{
            "description":"",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/TipoTarefa"
                }
              },
              "text/json": {
                "schema": {
                  "$ref":"#/components/schemas/TipoTarefa"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref":"#/components/schemas/TipoTarefa"
                }
              }
            }
          },
          "responses":{
            "401":{
              "description":"Unauthorized"
            },
            "400":{
              "description":"Usuario ja existe."
            },
            "200":{
              "description":"OK",
              "content":{
                "aplication/json":{
                  "schema":{
                    "type": "object",
                    "$ref":"#/components/schemas/TipoTarefa"                      
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas":{
        "jwt":{
          "type": "object",
          "properties": {
            "jwt":{
              "type":"string"
            }
          }
        },  
        "Retjwt":{
          "type": "object",
          "properties": {
            "status":{
              "type":"bollean"
            }
          }
        }, 
        "Return":{
          "type": "object",
          "properties": {
            "status":{
              "type":"bollean"
            },
            "message":{
              "type":"string"
            },
            "dados":{
              "type":"object"
            }
          }
        },   
        "SendMail":{
          "type": "object",
          "properties": {
            "to":{
              "type":"string"
            },
            "subject":{
              "type":"string"
            },
            "body":{
              "type":"string"
            }
          }
        },        
        "Users":{
          "type": "object",
          "properties": {
            "id":{
              "type":"int64"
            },
            "name":{
              "type":"string"
            },
            "email":{
              "type":"string"
            },
            "password":{
              "type":"string"
            }
          }
        },
        "Login":{
          "type": "object",
          "properties": {
            "email":{
              "type":"string"
            },
            "password":{
              "type":"string"
            }
          }
        },         
        "Tarefa":{
          "type": "object",
          "properties": {
            "idTarefa":{
              "type":"int64"
            },
            "idUsuario":{
              "type":"int64"
            },
            "nmTarefa":{
              "type":"string"
            },
            "dtTarefa":{
              "type":"datetime"
            },
            "dtLembreteTarefa":{
              "type":"datetime"
            },
            "nrRecorrenciaTarefa":{
              "type":"int"
            },
            "statusTarefa":{
              "type":"int"
            },            
            "Users":{
              "type": "object",
              "properties": {
                "id":{
                  "type":"int64"
                },
                "name":{
                  "type":"string"
                },
                "email":{
                  "type":"string"
                },
                "password":{
                  "type":"string"
                }
              }
            },
            "TipoTarefa":{
              "type": "object",
              "properties": {
                "idtar":{
                  "type":"int64"
                },
                "nametar":{
                  "type":"string"
                }
              }
            }
          }
        },
        "TipoTarefa":{
          "type": "object",
          "properties": {
            "idtar":{
              "type":"int64"
            },
            "nametar":{
              "type":"string"
            }
          }
        }
      },
      "securitySchemes":{
        "bearerAuth":{
          "type":"http",
          "scheme":"bearer",
          "bearerFormat": "JWT"
        }
      }
    }
  }

  